Decorations, Colliders, and Health Collectible
I added a few decorations which were mostly trees and metal boxes. Worked on colliders on Ruby and other Prefabs as well as the Tilemap. Finally, I worked on HealthCollectible item and made a new script for it. It has yet to be finished, but I did some good progress on it.